Fixed IDB. Moved packages dialog to own file.
